+++ crates/holochain-conductor-runtime-ffi/scripts/build.sh
@@ -6,7 +6,7 @@ cargo ndk --manifest-path ./Cargo.toml -t arm64-v8a -t x86 -t x86_64 \
  -o ../../libraries/holochain-service/src/main/jniLibs \
   build --release
 
-# Delete the types-ffi library from tauri-plugin-holochain-service, as it will be included via the dependency on holochain-service-client kotlin library
+# Delete the types-ffi library from tauri-plugin-holochain-service, as it will be included via the dependency on client kotlin library
 find "../../libraries/holochain-service/src/main/jniLibs/" -wholename '**/libholochain_conductor_runtime_types_ffi.so' -delete
 
 cargo run -p uniffi-bindgen-cli --release generate \