+++ crates/holochain-conductor-runtime-types-ffi/scripts/build-single-target.sh
@@ -5,10 +5,10 @@ TARGET=$1
 cargo build -p holochain-conductor-runtime-types-ffi
 
 cargo ndk --manifest-path ./Cargo.toml -t $TARGET \
-  -o ../../libraries/holochain-service-client/src/main/jniLibs \
+  -o ../../libraries/client/src/main/jniLibs \
   build --release
 
 cargo run -p uniffi-bindgen-cli --release generate \
   --library ../../target/$TARGET/release/libholochain_conductor_runtime_types_ffi.so \
-  --out-dir ../../libraries/holochain-service-client/src/main/java/ \
+  --out-dir ../../libraries/client/src/main/java/ \
   --language kotlin